+++ src/Groups/GroupDetails/index.js
@@ -1,1 +1,2 @@
-export { default } from './GroupDetailsContainer';
+export { default as GroupDetails } from './GroupDetails';
+export { default as GroupDetailsContainer } from './GroupDetailsContainer';