+++ src/Groups/GroupForm/index.js
@@ -1,1 +1,1 @@
-export { default } from './GroupForm';
+export { default as GroupForm } from './GroupForm';