+++ src/Groups/GroupsList/GroupsListFilters/index.js
@@ -1,0 +1,2 @@
+export { default as GroupsListFilters } from './GroupsListFilters';
+export { default as GroupsListFiltersContainer } from './GroupsListFiltersContainer';