+++ src/common/const/routes.js
@@ -1,6 +1,4 @@
 export const GROUPS_ROUTE = '/finance/groups';
-export const GROUP_VIEW_ROUTE = `${GROUPS_ROUTE}/view/`;
-export const GROUP_EDIT_ROUTE = `${GROUPS_ROUTE}/edit/`;
 export const FISCAL_YEAR_ROUTE = '/finance/fiscalyear';
 export const FISCAL_YEAR_VIEW_ROUTE = `${FISCAL_YEAR_ROUTE}/view/`;
 export const FISCAL_YEAR_EDIT_ROUTE = `${FISCAL_YEAR_ROUTE}/edit/`;