+++ test/bigtest/tests/groups/GroupDetails/GroupSummaryDetails-test.js
@@ -1,7 +1,7 @@
 import { beforeEach, describe, it } from '@bigtest/mocha';
 import { expect } from 'chai';
 
-import { GROUP_VIEW_ROUTE } from '../../../../../src/common/const';
+import { GROUPS_ROUTE } from '../../../../../src/common/const';
 
 import setupApplication from '../../../helpers/setup-application';
 import GroupDetailsInteractor from '../../../interactors/groups/GroupDetailsInteractor';
@@ -17,7 +17,7 @@ describe('Group summary details', () => {
     this.server.create('fiscalYear');
     this.server.createList('groupSummary', 10);
 
-    this.visit(`${GROUP_VIEW_ROUTE}${group.id}`);
+    this.visit(`${GROUPS_ROUTE}/${group.id}/view`);
     await groupDetails.whenLoaded();
   });